simplifying refactor
